% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatellDS.R
\name{calculatellDS}
\alias{calculatellDS}
\title{Partial Model Predictions}
\usage{
calculatellDS(y, pi)
}
\arguments{
\item{y}{is independent variable for a logistic model.}

\item{pi}{ia a vector with predictions values for each observation.}
}
\value{
a vector or matrix of predictions.
}
\description{
Partial computation of prediction for logistic regression models.
}
\details{
This function already needs some tests and improvements.
}
\author{
Paula R. Costa e Silva
}
