% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPi.R
\name{findPi}
\alias{findPi}
\title{Predictions on Each Observation}
\usage{
findPi(bind.x, beta, family)
}
\arguments{
\item{bind.x}{is a data matrix of independent variables.}

\item{beta}{is a list of regression coefficients (estimated betas).}

\item{family}{a string character with the name of the link function to be used in the analysis.
If \code{family} is set to 'binomial', a logit function is applied.
If \code{family} is set to 'poisson', a log function applied.}
}
\value{
a vector of predictions on each observation.
}
\description{
Computes the value of predictions on each observation (P(i)), given bind.x(input) and estimated betas.
}
\details{
Internal function.
Computes the predictions on each observation by a logit function (used for logistic regression)
or log function (used for poisson regression).
}
\author{
Paula R. Costa e Silva
}
