% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findW.R
\name{findW}
\alias{findW}
\title{Matrix W}
\usage{
findW(pi, family)
}
\arguments{
\item{pi}{is a list of predictions on each observation.}

\item{family}{a character that determines the likelihood function.
If \code{family} is set to 'binomial', a likelihood for binary logistic regression is applied
If \code{family} is set to 'poisson', a likelihood for poisson regression is applied.}
}
\value{
the W matrix
}
\description{
Calculate the matrix W with all diagonal values as p(i).
}
\details{
It is a partial computation for Hessian Matrix (second derivative).
If the Hessian Matrix is \eqn{H=XWX^T}, the W matrix is \eqn{W = P_i(1-P_i)}.
}
\author{
Paula R. Costa e Silva
}
