% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnova.R
\name{getAnova}
\alias{getAnova}
\title{Anova}
\usage{
getAnova(beta, formula, media.y)
}
\arguments{
\item{beta}{is a list of the regression coefficients.}

\item{formula}{a string character to be transformed as an object of class formula.}

\item{media.y}{a numeric value, the arithmetic mean for a given response variable.}
}
\value{
A list with components:
\item{rows.x}{the number of rows present in x.}
\item{rows.y}{the number of rows present in y.}
\item{cols.x}{the number of cols present in x.}
\item{cols.y}{the number of cols present in y.}
\item{residuals}{the prediction error.}
\item{sst}{total sum of squares.}
\item{sse}{sum of squares error.}
\item{lbx}{the number of study variables.}
\item{quad.y}{sum of total squares for response variable.}
\item{y.hat}{the fitted values.}
}
\description{
Partial computations for Anova analysis.
}
\details{
Computes the partial equations for test regression models.
}
\author{
Paula Raissa Silva
}
