% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCondSamplesCounting.R
\name{getCondSamplesCounting}
\alias{getCondSamplesCounting}
\title{Conditional Samples Counting}
\usage{
getCondSamplesCounting(varName, catName, cat, xREF, datasources)
}
\arguments{
\item{varName}{a study variable, it must be a numeric vector.}

\item{catName}{a cetegorical variable.}

\item{cat}{a character, given category to compute the sample size.}

\item{xREF}{a numeric value, given reference do calculate the sample size.}

\item{datasources}{a list of data connection parameters.}
}
\value{
The numeric value corresponding to a percentile.
}
\description{
Computes the x value corresponding to a given percentile and categories.
}
\details{
Returns the x value corresponding to a given percentile \link{0..1}
}
\section{Dependencies}{

\code{\link{getVarByName}}
}

\author{
Rui Camacho, Paula Raissa
}
