% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGm.R
\name{getGm}
\alias{getGm}
\title{getGm}
\usage{
getGm(x, na.rm = TRUE, zero.propagate = FALSE)
}
\arguments{
\item{x}{is name of a variable}

\item{na.rm}{logical. Should misssing values (including NaN) be removed?}

\item{zero.propagate}{logical. Control the propagation of zeros.}
}
\value{
the sum of logs
}
\description{
getGm returns the sum of logs of all values present in its arqguments
}
\details{
The geometric mean is one type of average.
It differs to arithmetic mean when the arithmetic mean adds items and the geometric mean multiplies items.
Formally the geometric mean is the i_n root of the \emph{n} numbers products or \emph{exp} to the mean log of \emph{x}.
}
\section{Dependencies}{

\code{\link{getVarByName}}
}

\author{
Paula R. Costa e Silva
}
