% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKurtosis.R
\name{getKurtosis}
\alias{getKurtosis}
\title{Kurtosis}
\usage{
getKurtosis(x, mean)
}
\arguments{
\item{x}{is a study variable.}

\item{mean}{the arithmetic mean.}
}
\value{
A list with components:
\item{m2}{second sample moment.}
\item{m4}{fourth sample moment.}
}
\description{
Kurtosis partial computation for a study variable.
}
\details{
Partial calculations of skew measure.

Considering x\link{i} for elements of x and x* for their mean, so \eqn{m[r] = \sum[i](x[i]-x*)^r/n} is a partial sample moments of order r.

This function computes the 2nd and 4th partial sample moments.
}
\section{Dependencies}{

\code{\link{getVarByName}}
}

\author{
Paula R. Costa e Silva
}
