% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLength.R
\name{getLength}
\alias{getLength}
\title{Length of an Object}
\usage{
getLength(x)
}
\arguments{
\item{x}{is an R object.}
}
\value{
a numerical value.
}
\description{
Computes the size of a given vector.
}
\details{
Get or set the length of vectors (including lists) and factors, and of any other R object for
which a method has been defined.

For vectors (including lists) and factors the length is the number of elements.
For an environment it is the number of objects in the environment, and NULL has length 0.
For expressions and pairlists (including language objects and dotlists) it is the length of the pairlist chain.
All other objects (including functions) have length one.
}
\section{Dependencies}{

\code{\link{getVarByName}}
}

\author{
Paula R. Costa e Silva
}
