% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevels.R
\name{getLevels}
\alias{getLevels}
\title{Compute the levels for a categorical variable.}
\usage{
getLevels(x, g, datasources = NULL)
}
\arguments{
\item{x}{is a study variable.}

\item{g}{is the categorical variable.}
}
\value{
a vector of fitted values.
}
\description{
Calculates the fitted values in each data node.
}
\details{
Considering \code{y} as a response variable and x as study variable, the fitted values are the y-values that
would expect for the given x-values according to the best-fitting straight line.
}
\section{Dependencies}{

\code{\link{getVarbyFormula}}
}

\author{
Paula R. Costa e Silva
}
