% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMax.R
\name{getMax}
\alias{getMax}
\title{Maximum}
\usage{
getMax(x)
}
\arguments{
\item{x}{is a study variable.}
}
\value{
The maximum numeric value inside the vector.
}
\description{
Computes the maximum value in the sample according to a study variable.
}
\details{
Return the maximum of all the values present in their arguments, as integer if all are logical or integer,
as double if all are numeric, and character otherwise.

The maximum of a numeric empty set is +Inf which ensures transitivity.
For numeric x ds.max(x) == -Inf whenever ds.length(x) == 0 (after removing missing values if requested).

By definition the max of a numeric vector containing an NaN is NaN, except that the max of any vector containing
an NA is NA even if it also contains an NaN. Note that ds.max(NA, Inf) == NA even though the maximum would be Inf whatever
the missing value actually is.

Character versions are sorted lexicographically. The max of an empty character vector is defined to be character NA.
}
\section{Dependencies}{

\code{\link{getVarByName}}
}

\author{
Paula R. Costa e Silva
}
