% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMean.R
\name{getMean}
\alias{getMean}
\title{Arithmetic Mean}
\usage{
getMean(x)
}
\arguments{
\item{x}{is a study variable.}
}
\value{
The arithmetic mean of the values in x.
}
\description{
Computes the arithmetic mean.
}
\details{
Generic function for the arithmetic mean.
}
\section{Dependencies}{

\code{\link{getVarByName}}
}

\author{
Paula R. Costa e Silva
}
