% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeCDF.R
\name{getNodeCDF}
\alias{getNodeCDF}
\title{Get CDf for Node}
\usage{
getNodeCDF(varName, datasources)
}
\arguments{
\item{varName}{a character, the name of study variable.}

\item{datasources}{a list of parameters to access files sytems or databases.}
}
\description{
Computes the x value from given percentile.
}
\details{
Read each data file (node) and compute the local cumDistFunc for each data node.
}
\author{
Rui Camacho, Paula Raissa
}
