% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumNa.R
\name{getNumNa}
\alias{getNumNa}
\title{Count of missing values}
\usage{
getNumNa(x)
}
\arguments{
\item{x}{is a numerical variable.}
}
\value{
the NA count as numerical value.
}
\description{
Computes the missing values of a given vector.
}
\details{
Counts all na or another missing value indicator into a vector.
}
\section{Dependencies}{

\code{\link{getVarByName}}
}

\author{
Paula R. Costa e Silva
}
