% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPercentile.R
\name{getPercentile}
\alias{getPercentile}
\title{Get Conditional Percentile}
\usage{
getPercentile(samplesList)
}
\arguments{
\item{samplesList}{a vector,  the local cumDistFunc for each data node.}
}
\value{
Combined percentile value.
}
\description{
Computes the conditional percentile according to the sample list.
}
\details{
Get the percentile for given x value and cdf (sample list).
}
\author{
Rui Camacho, Paula Raissa
}
