% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProd.R
\name{getProd}
\alias{getProd}
\title{Product of Study Variable Elements}
\usage{
getProd(x, naRm = FALSE)
}
\arguments{
\item{x}{is the name of a study variable.}

\item{na.rm}{logical (TRUE/FALSE). Remove, or not the missing values.}
}
\value{
The product as a numerical value.
}
\description{
It computes the product of all values in its arguments.
}
\details{
If na.rm is FALSE an NA value in any of the arguments will cause a value of NA to be returned,
otherwise NA values are ignored.

Logical true values are regarded as one, false values as zero. For historical reasons, NULL is accepted and
treated as if it were numeric(0).
}
\section{Dependencies}{

\code{\link{getVarByName}}
}

\author{
Paula R. Costa e Silva
}
