% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSkewness.R
\name{getSkewness}
\alias{getSkewness}
\title{Skewness}
\usage{
getSkewness(x, mean)
}
\arguments{
\item{x}{is a study variable}

\item{mean}{the arithmetic mean as a numeric value}

\item{datasources}{a list of data connection parameters.}
}
\value{
A list with components:
\item{m2}{second sample moment.}
\item{m3}{third sample moment.}
}
\description{
Skewness partial computation for a study variable.
}
\details{
Partial calculations of skew measure.

Considering \eqn{x[i]} for elements of \eqn{x} and \eqn{x*} for their mean, so \eqn{m[r] = \sum{i}(x[i]-x*)^r}
is a partial sample moments of order r.

This function computes the 2nd and 3th partial sample moments.
}
\section{Dependencies}{

\code{\link{getVarByName}}
}

\author{
Paula R. Costa e Silva
}
