% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSumOfSquares.R
\name{getSquaredSum}
\alias{getSquaredSum}
\title{Sum of Squares}
\usage{
getSquaredSum(x)
}
\arguments{
\item{x}{is a study variable.}
}
\value{
the sum of squares as a numeric value.
}
\description{
Computes the sum of squares for a study variable.
}
\details{
Requires the function \code{\link{getVarByName}}.
}
\author{
Paula R. Costa e Silva
}
