% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariance.R
\name{getVariance}
\alias{getVariance}
\title{Sample Variance}
\usage{
getVariance(x, mean)
}
\arguments{
\item{x}{a character, the name of a study variable.}

\item{mean}{numerical, the arithmetic mean.}
}
\value{
the variance.
}
\description{
Computes the variance for a study variable.
}
\details{
By default, computes the sample variance.
}
\author{
Paula R. Costa e Silva
}
